moving hashing of vec outside of generate probes

+++ Repetition/src/main/scala-2.12/hashfunctions/HashFunction.scala
@@ -2,7 +2,7 @@ package hashfunctions
 
 trait HashFunction {
   def apply(v: Array[Float]): Array[Int]
-  def generateProbes(v : Array[Float]): Array[Array[Int]]
+  def generateProbes(v : Array[Int]): Array[Array[Int]]
 }
